add test classes for HarperDBBase and HarperDBError
